Absorb buildApp into modular-scripts
